more priors into model file
